Move utilities into rootDir
